fix: rework response archiver lambda
