+++ .gitignore
@@ -6,7 +6,7 @@
 *.tfstate.*
 
 # Terraform lock file
-.bootstrap/.terraform.lock.hcl
+.terraform.lock.hcl
 
 # Terragrunt
 .terragrunt-cache
