+++ aws/lambdas/code/response_archiver/package.json
@@ -1,7 +1,7 @@
 {
   "name": "archive-form-responses",
   "version": "1.0.0",
-  "main": "archiver.js",
+  "main": "response_archiver.js",
   "type" : "module",
   "license": "MIT",
   "scripts": {
