+++ aws/lambdas/response_archiver.tf
@@ -22,11 +22,11 @@ resource "aws_lambda_function" "response_archiver" {
   s3_object_version = aws_s3_object.response_archiver_code.version_id
   function_name     = "Response_Archiver"
   role              = aws_iam_role.lambda.arn
-  handler           = "archiver.handler"
+  handler           = "response_archiver.handler"
 
   source_code_hash = data.archive_file.response_archiver_code.output_base64sha256
   runtime          = "nodejs18.x"
-  timeout          = 10
+  timeout          = 300
 
   environment {
     variables = {
